Move response&obs logic to own folder
