+++ src/ert/config/responses/general_observation.py
