+++ src/ert/config/responses/response_properties.py
