+++ src/ert/storage/local_ensemble.py
@@ -28,7 +28,7 @@ from pydantic import BaseModel
 from typing_extensions import deprecated
 
 from ert.config.gen_kw_config import GenKwConfig
-from ert.config.observations import ObservationsIndices
+from ert.config.responses.observations import ObservationsIndices
 from ert.storage.mode import BaseMode, Mode, require_write
 
 from ..config import GenDataConfig, ResponseTypes
